Expose the setup, setup_gpio and setup_sys
This is a backwards-incompatible change while preparing for wiringPi
version 2 support.  This commit still supports wiringPi version 1.

Previously the setup function - wiringPiSetup() - was called when the
NIF was loaded.  The downside with that approach was that wpi had to
be run as root, although the nice thing was that a user wouldn't have
to worry about calling setup before accessing pins.

After some consideration I've (despite breaking backwards
compatibility) decided to remove the setup from the NIF loading and
let the user choose the method.

Hopefully fixes issue #4.